Move QuantizationScheme to burn-tensor
