+++ crates/burn-core/src/module/mod.rs
@@ -1,7 +1,9 @@
 mod base;
 mod display;
 mod param;
+mod quantize;
 
 pub use base::*;
 pub use display::*;
 pub use param::*;
+pub use quantize::*;
