+++ crates/burn-tensor/src/tensor/quantization/scheme.rs
