+++ crates/burn-tensor/src/tensor/quantization/strategy.rs
